refactor(bhwi): isolate Coldcard common adapters
- keep multisig commands in protocol-native address formats
- move common command and response mappings to the adapter layer
- retain descriptor validation at the common boundary
